app/requirements.txt
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
pymongo==4.6.0

# Blockchain
web3==6.11.3
eth-account==0.10.0
eth-utils==2.3.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
httpx==0.25.2
requests==2.31.0